pyproject.toml
[tool.poetry]
name = "sketchbench-data-ingestion-tester"
version = "0.1.0"
description = "Fake data producer for testing Kafka data ingestion"
authors = ["Frank Blechschmidt <frankb2@illinois.edu>"]
license = "Apache-2.0"
packages = [
    { include = "main.py" },
]

[tool.poetry.dependencies]
python = "^3.7"
Faker = "^8.9.0"
click = "^8.0.1"
click-pathlib = "^2020.3.13"
loguru = "^0.5.3"
orjson = "^3.5.4"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
bandit = "^1.7.0"
dodgy = "^0.2.1"
mypy = "^0.910"
safety = "^1.10.3"
wemake-python-styleguide = "^0.15.3"
types-orjson = "^0.1.1"

[tool.poetry.scripts]
sketchbench-data-ingestion-tester = "main:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
